requirements.txt
streamlit
pandas
fpdf2
Pillow
openpyxl
num2words
